Documentando API no swagger e realizando tratamento de exceções. Foi tratada a exceção *Delete* Para Client encontrado e erro para Client nao encontrado.

+++ src/main/java/com/example/itauchallenge/service/ClientService.java
@@ -2,6 +2,7 @@ package com.example.itauchallenge.service;
 
 import com.example.itauchallenge.dto.ClientDTO;
 import com.example.itauchallenge.entity.Client;
+import com.example.itauchallenge.exception.ClienteNotFoundException;
 import com.example.itauchallenge.repository.ClientRepository;
 import org.springframework.stereotype.Service;
 
@@ -24,40 +25,35 @@ public class ClientService {
         client.setFirstName(clientDTO.firstName());
         client.setLastName(clientDTO.lastName());
         client.setParticipation(clientDTO.participation());
-        if (client == null
-                || client.getFirstName() == null
-                || client.getLastName() == null
-                || client.getParticipation() == null){
-            throw new NullPointerException();
-        }
         return clientRepository.save(client);
-
     }
 
     public List<Client> listAllClients() {
-        List<Client> allClients = clientRepository.findAll();
-        if (allClients == null || allClients.isEmpty()){
-            throw new NullPointerException();
-        }
-        return allClients;
+        return clientRepository.findAll();
     }
 
-    public void delete(Long id) {
-        clientRepository.deleteById(id);
+    public String delete(Long id) {
+        Optional<Client> deleteClient = clientRepository.findById(id);
+
+        if (deleteClient.isPresent()) {
+            clientRepository.deleteById(id);
+            return "Usuário deletado com sucesso";
+        } else {
+            throw new ClienteNotFoundException("Cliente não encontrado com o ID: " + id);
+        }
     }
 
     public Client update(Long id, ClientDTO clientDTO){
         Optional<Client> optionalClient = clientRepository.findById(id);
         if (optionalClient.isPresent()){
             Client client = optionalClient.get();
-
             client.setFirstName(clientDTO.firstName());
             client.setLastName(clientDTO.lastName());
             client.setParticipation(clientDTO.participation());
 
             return clientRepository.save(client);
         } else {
-            throw new RuntimeException("Client not found with ID: " + id);
+            throw new RuntimeException("Cliente não encontrando com o ID: " + id);
         }
     }
 }
